Refactor Problem module name
