+++ rg_instructor_analytics/views/Suggestion.py
@@ -11,7 +11,7 @@ from course_modes.models import CourseMode
 from django_comment_client.utils import JsonResponse
 from rg_instructor_analytics.utils.AccessMixin import AccessMixin
 from rg_instructor_analytics.views.Funnel import GradeFunnelView
-from rg_instructor_analytics.views.Problem import ProblemHomeWorkStatisticView
+from rg_instructor_analytics.views.problem import ProblemHomeWorkStatisticView
 
 
 class BaseSuggestion(object):
